back-end/requirements.txt
Flask==3.0.2
requests==2.31.0
python-docx==1.1.0
PyPDF2==3.0.1
numpy
